Make Ant move randomly
Ants can move to any cell, including in diagonal, but the cell is only
valid if it is the direct neighbor of a cube or on the floor.

We randomly chose which cell to move to.

Thanks to testing the adjacent_positions method is fixed, tests were
passing likely because we were using test position of (0,0,0).

Tests:
- test_ant.py OK