Change generator to list to reuse it
Because the list of direct neighbors was provided as a generator it
would only work for the first pheromone we were iterating on. Leaving
all the others at 0 and never being chosen.

Replacing it with a list allows us to reuse it and handle as many
pheromones as we want.

Tests:
- python test_ant.py