Add script to import sculpture in Blender
Read an export file generated by Ants and create cubes for each position
in the env that has pheromones (no matter which ones)

The cubes are then joined into a single piece, this will allow us to
actually print it.

Path of file to read is hardcoded for now.

Tests:
- multiple run of that script from within blender created sculpture!